Implement plural inflections feature and simplify code
Optimizations were made to how inflections are handled in the application. A modal was created where users can choose to include plural forms in the inflections. This update enhances user experience by giving them control over their inflection preferences. In line with this, corresponding changes were made to the 'getInflections' and 'getUpdatedAliases' methods.

Some event listener code related to mouse clicks and intervals were commented out as they were unnecessary given the application's current requirements.

Furthermore, a 'httpGetMorpher' method was added to clean up the process of making HTTP requests. The method 'httpGetMorpherStub' is a stub of 'httpGetMorpher', it was added to assist in the testing of the application.