Refactor settings code into separate file
Moved the code related to plugin settings from main.ts file to a newly created settings.ts file. The main reason for this modification was to improve code organization, making it easier to manage and test. Now, all the settings related code, including interfaces, default settings, and the setting class, resides in a designated settings file.