Refactor AddAliasesModal code into separate file
Moved the code related to AddAliasesModal from the main.ts file to a newly created add_aliases_modal.ts file. This was done to improve the code's readability and maintainability. Now, all the modal related code is in its own designated file, creating a modular structure which enhances the organization of the code.