Update README.md, modify labels in add_aliases_modal.ts and settings.ts
Completely overhauled README.md to provide detailed information about the new usage of the Alias Inflector Plugin. This includes the ability of a single command to add all possible inflections of a word or phrase as aliases, and usage of online inflection service morpher.ru. Changed labels in add_aliases_modal.ts and settings.ts to be more descriptive. This provides clarity to users about what each option does in the GUI.