Refactor AddAliasesModal class with InflectionOptions
The AddAliasesModal class's constructor was refactored to take an InflectionOptions object instead of two separate boolean flags. This change makes creation of AddAliasesModal objects more manageable and extendable if more options are added in the future. It also touched several methods in other classes to conform to this change. InflectionOptions interface was created to encapsulate the two flags, improving code structure and clarity.