Update eslint config
A new script for ESLint has been added to package.json. This change helps to check the TypeScript files for any linting issues. The ESLint configuration file (.eslintrc.yml) was also updated to ignore certain paths. Additionally, instructions for development, as well as how to run tests and linters, were added to the README file.